jarvis-assistant/requirements.txt
# Core AI and Orchestration
ollama>=0.1.7
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.2
psutil>=5.9.8

# Voice (STT & Wake Word)
SpeechRecognition>=3.10.1
PyAudio>=0.2.14
pvporcupine>=3.0.0

# Voice (Speech Synthesis - Kokoro TTS)
kokoro-onnx>=0.2.0
sounddevice>=0.4.6

# Computer Vision
opencv-python>=4.9.0.80

# Skills and Search
duckduckgo-search>=5.3.0
wikipedia>=1.4.0
spotipy>=2.23.0
pywhatkit>=5.4
pyautogui>=0.9.54

# Data Processing and Utilities
numpy>=1.26.4
Pillow>=10.2.0
scipy>=1.12.0
requests>=2.31.0
scikit-learn>=1.4.0

